Added BIOParallel sub-hierarchy
